backend/requirements.txt
fastapi
uvicorn
aiofiles
python-multipart
pydantic
python-dotenv
redis>=5.0.0
openai
pypdf